Use generators to lift the callback out of state::fetch.
The thinking here was to do something similar in the command processor
to get rid of basically all the callbacks, which would let the command
processor remain "synchronous", as well as state::fetch.

But it really doesn't solve anything. Yielded items can't borrow from
within the generator, and limitations on the type system make lifetimes
generally harder to manage. Because the command processor ties many
commands together, and each generator would have a different type, there
would also be an explosion of boilerplate to forward across types.

At the end of the day, callbacks still have simpler lifetimes, and
pushing async into state::fetch wouldn't be that bad. (It's already
bending over backwards relative to the rest of the state API anyway.)
Further, async can be added from the bottom up: a sync FetchReceiver can
trivially meet the async interface, and then the same can be done with
SendResponse, only converting the server itself at the end.